requirements.txt
python-dotenv==1.0.1
fastapi==0.112.2
uvicorn==0.30.6
openai==1.42.0
aiohttp==3.10.5
asyncpg==0.29.0
pytz==2024.2